Revert "Add pidfile support"
This reverts commit d4b330b1584897f63cbfc40afef4fe79f2dc5db0.  Derp,
we can't access a file by name once we are chrooted.  nor can we
delete a file owned by root once we change uids.  There doesn't really
seem to be a good reason to have a pid file, the user program will
communicate with the daemon through a socket and no one else really
needs to talk to the daemon...